Enter search mode when clicking pathbar.

+++ editor/large-pathbar.vala
@@ -64,6 +64,11 @@ private class LargePathbar : Box, Pathbar
             return;
 
         _set_path (type, path);
+        if (invisible_button != null)
+        {
+            Variant variant = new Variant.string (complete_path);
+            ((!) invisible_button).set_detailed_action_name ("ui.open-search(" + variant.print (false) + ")");
+        }
         update_active_button_cursor (type, ref active_button);
     }
     private void _set_path (ViewType type, string path)
@@ -79,6 +84,13 @@ private class LargePathbar : Box, Pathbar
         bool destroy_all = false;
         bool maintain_all = false;
         @foreach ((child) => {
+                if (child is InvisibleButton)
+                {
+                    child.destroy ();
+                    invisible_button = null;
+                    return;
+                }
+
                 if (child is Label)
                 {
                     if (destroy_all)
@@ -138,6 +150,7 @@ private class LargePathbar : Box, Pathbar
             }
         }
 
+        add_invisible_button ();
         @foreach ((child) => child.show ());
     }
     private static inline void update_config_style_class (bool type_is_config, StyleContext context)
@@ -222,6 +235,14 @@ private class LargePathbar : Box, Pathbar
     * * widgets management
     \*/
 
+    InvisibleButton? invisible_button = null;
+
+    private void add_invisible_button ()
+    {
+        invisible_button = new InvisibleButton ();
+        add ((!) invisible_button);
+    }
+
     private void add_slash_label ()
     {
         add (new Label ("/"));
@@ -279,6 +300,20 @@ private class LargePathbar : Box, Pathbar
             item.set_detailed_action_name (item.default_action);
             item.get_style_context ().remove_class ("active");
         }
+    }
+}
+
+private class InvisibleButton : Button
+{
+    construct
+    {
+        set_detailed_action_name ("ui.open-search('/')");
+        hexpand = true;
+        vexpand = true;
+        can_focus = false;
+        focus_on_click = false;
+        StyleContext context = get_style_context ();
+        context.add_class ("invisible");
     }
 }
